requirements.txt
aiofiles==24.1.0
aiogram==3.19.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
alembic==1.14.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asyncpg==0.30.0
attrs==25.1.0
bcrypt==4.3.0
billiard==4.2.1
black==25.1.0
build==1.2.2.post1
CacheControl==0.14.2
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cleo==2.1.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
crashtest==0.4.1
distlib==0.3.9
dnspython==2.7.0
dulwich==0.22.7
email_validator==2.2.0
et_xmlfile==2.0.0
fastapi==0.115.8
fastapi-cache2==0.2.2
fastapi-cli==0.0.7
fastjsonschema==2.21.1
filelock==3.17.0
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
installer==0.7.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.5
keyring==25.6.0
kombu==5.5.0
magic-filter==1.0.12
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.6.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==2.2.2
openpyxl==3.1.5
orjson==3.10.15
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pathspec==0.12.1
pendulum==3.0.0
pillow==11.1.0
pkginfo==1.12.0
platformdirs==4.3.6
pluggy==1.5.0
poetry==2.0.1
poetry-core==2.0.1
prompt_toolkit==3.0.50
propcache==0.3.0
pycparser==2.22
pydantic==2.10.6
pydantic-extra-types==2.10.2
pydantic-settings==2.8.1
pydantic_core==2.27.2
pygame==2.6.1
Pygments==2.19.1
PyJWT==2.10.1
pyproject_hooks==1.2.0
pytest==8.3.5
pytest-asyncio==0.25.3
pytest-dotenv==0.5.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
RapidFuzz==3.12.1
redis==5.2.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rich-toolkit==0.13.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
time-machine==2.16.0
tomlkit==0.13.2
trove-classifiers==2025.1.15.22
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
vine==5.1.0
virtualenv==20.29.1
watchfiles==1.0.4
wcwidth==0.2.13
websockets==15.0
xattr==1.1.4
yarl==1.18.3
